backend/requirements.txt
Django==5.1.1
psycopg[binary]==3.2.1
django-environ==0.11.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
celery==5.4.0
redis==5.0.8
gunicorn==22.0.0
python-dateutil==2.9.0.post0
Pillow
django-crontab==0.7.1
